requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
boto3==1.40.61
botocore==1.40.61
click==8.3.1
fastapi==0.128.0
frozenlist==1.8.0
h11==0.16.0
idna==3.11
jmespath==1.1.0
multidict==6.7.1
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079150/work
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-ulid==3.1.0
s3transfer==0.14.0
setuptools==80.10.1
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wheel==0.46.3
wrapt==1.17.3
yarl==1.22.0
